pyproject.toml
[project]
name = "bastion-session"
version = "0.1.0"
description = "Bastion session manager CLI"
requires-python = ">=3.10"
authors = [
  { name = "Remote Workstation Team" }
]
readme = "README.md"
dependencies = [
  "click>=8.1",
  "rich>=13.7",
  "oci>=2.133.1",
  "python-dateutil>=2.8",
  "typing-extensions>=4.8"
]

[project.scripts]
bastion-session = "bastion_session_cli.main:cli"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
